+++ src/overlay/aria.js
@@ -1,0 +1,15 @@
+/**
+ * http://www.w3.org/TR/wai-aria-practices/#trap_focus
+ * @author:yiminghe@gmail.com
+ */
+KISSY.add("overlay/aria", function() {
+    function Aria() {
+    }
+
+    Aria.ATTRS = {
+        aria:{
+            view:true
+        }
+    };
+    return Aria;
+});