+++ src/overlay/overlay.js
@@ -2,22 +2,22 @@
  * model and control for overlay
  * @author:yiminghe@gmail.com
  */
-KISSY.add("overlay/overlay", function(S, UIBase, Component, OverlayRender) {
+KISSY.add("overlay/overlay", function(S, UIBase, Component, OverlayRender, Aria) {
     function require(s) {
         return S.require("uibase/" + s);
     }
 
-    var Overlay= UIBase.create(Component.ModelControl, [
+    var Overlay = UIBase.create(Component.ModelControl, [
         require("contentbox"),
         require("position"),
         require("loading"),
         require("align"),
         require("resize"),
-        require("mask")]);
+        require("mask"),Aria]);
 
-    Overlay.DefaultRender=OverlayRender;
+    Overlay.DefaultRender = OverlayRender;
 
     return Overlay;
 }, {
-    requires:['uibase','component','./overlayrender']
+    requires:['uibase','component','./overlayrender','./aria']
 });